requirements/prod.txt
Django==4.2
django-cleanup==8.0.0
django-phonenumber-field[phonenumbers]==7.2.0
django-widget-tweaks==1.5.0
pyjwt==2.8.0
pillow==10.1.0
python-dotenv==1.0.0
sorl-thumbnail==12.10.0
